refactor(client): [CLI-026] :recycle: Change Import's names
