+++ client/core/src/main/java/fr/esgi/pa/alliodesk/core/request/WorkspaceRequest.java
@@ -1,8 +1,8 @@
 package fr.esgi.pa.alliodesk.core.request;
 
 import com.google.gson.Gson;
-import fr.esgi.pa.alliodesk.core.Service;
-import fr.esgi.pa.alliodesk.core.Workspace;
+import fr.esgi.pa.alliodesk.core.models.Service;
+import fr.esgi.pa.alliodesk.core.models.Workspace;
 import fr.esgi.pa.alliodesk.core.form.InfoInForm;
 import org.apache.http.client.methods.*;
 import org.apache.http.util.EntityUtils;
